Change the namespace of the Presenter classes
