--- app/presenters/shopify/variant_presenter.rb
@@ -1,11 +1,0 @@
-module Shopify
-  class VariantPresenter < Delegator
-    def weight_unit
-      'oz'
-    end
-
-    def default_pos_image
-      images.first
-    end
-  end
-end
